Fixed the DeleteButton
I took out the noItemsFound prop being passed to DeleteButton, abd replaced it with the disabled prop, which directly accesses the state. This definitely works, and is simpler, though I would like to know why I couldn't pass the noItemsFound function. I think the answer is that I wasn't calling it properly.